requirements.txt
# 核心框架
llama-index-core==0.13.0
llama-index-llms-ollama==0.7.0
llama-index-llms-zhipuai==0.4.0
llama-index-workflows==1.3.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index>=0.9.0
chromadb>=0.4.0
chroma-hnswlib>=0.7.0

# 文本分块相关
langchain>=0.1.0
langchain-text-splitters>=0.1.0

# 文档处理
PyMuPDF>=1.23.0
markdown-it-py>=3.0.0
markitdown>=0.1.0
pandas==2.2.3
openpyxl>=3.1.0
xlrd>=2.0.0
pypdf==5.9.0
beautifulsoup4==4.13.4

# 文本处理和向量化
sentence-transformers>=2.2.0
transformers==4.55.0
torch==2.8.0
numpy==2.3.2

# 异步处理
asyncio
aiofiles>=23.0.0
nest-asyncio==1.6.0

# 配置和日志
pyyaml>=6.0
python-dotenv==1.1.1
requests==2.32.4
tiktoken==0.10.0
pydantic==2.11.7

# 工具库
tqdm==4.67.1
pathlib
typing-extensions>=4.5.0

# Google API (组员项目需要)
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2

# OpenAI (组员项目需要)
openai==1.99.1

# Ollama (组员项目需要)
ollama==0.5.2

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0

# 开发工具
black>=23.0.0
flake8>=6.0.0